requirements.txt
lyricsgenius
python-dotenv